Fixing errors before making changes
The main menu is now linked to the first scene.
The game scene has been renamed to Vampire Runner Scene instead of VR.
The 'GameContain' reference in the MenuScript has been given an Empty object as it was blank (I'm guessing this is where Heidi intended the game to be, inside this gamecontain object'
Updated to version latest unity version, we need to work out a version that we will all use!