Player Mechanics
I have made it so the player can now slide for x seconds (and have a reload on when they can next slide for y seconds). Jump and slide animations that player correctly. 
Death animation when the palyer collides with a hazard. (also the player stops moving)
Added Gravestone, Spikes and Crossbow Bolt hazards to the game (they are tagged as 'hazard').
Added a DevTestingObj which allows you to restart by pressing R.
Changed the jump and sliding inputs to X and Z and changed a few things around in the player controller script to make it better hopefully.
Player animation controller and animations.